Ignore link refs and link ref definitions
Links that look like `[#45][#45]` will be converted incorrectly because
it does not match the existing boundary. Specifically, the *second*
instance of `[#45]` will be converted into an inline link, even though
this is a link reference. Similarly, the issue/pull linking would
overconvert the definition of link references (`[foo]: uri`).

This change specifically looks for these two cases (link ref and link
ref definition) in the pre/post match values from the `#gsub` match
and if either of these cases is met, the original matching text is
returned unmodified.

It is *possible* that this could be partially fixed in the main regex
definitions, but those are already complex and while I understand regex
well, I could not figure out the right way to manage this because of the
contextual complexity involved.

This would be much better managed as transforms over a parsed Markdown
AST where the replacements could be applied to pure text nodes while
already ignoring things that are identified as link definitions (maybe
[comrak](https://crates.io/crates/comrak)).

Resolves: #3